clean up google lens ocr
